+++ dbldatagen/utils.py
@@ -236,7 +236,7 @@ def split_list_matching_condition(lst, cond):
      ['id'], ['city_id', 'city_pop', 'city_id', 'city_pop'], ['id']]`
 
     :arg lst: list of items to perform condition matches against
-    :arg cond: lambda function or function taking single argument and returning True or Fals
+    :arg cond: lambda function or function taking single argument and returning True or False
     :returns: list of sublists
     """