+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to the Databricks Labs Data Generator will be documented in 
 #### Changed
 * Adjusted column build phase separation (i.e which select statement is used to build columns) so that a 
   column with a SQL expression can refer to previously created columns without use of a `baseColumn` attribute
+* Changed build labelling to comply with PEP440
 
 #### Fixed 
 
@@ -15,7 +16,7 @@ All notable changes to the Databricks Labs Data Generator will be documented in 
 * Parsing of SQL expressions to determine column dependencies
 
 #### Notes
-* This does not change actual order of column building - but adjusts which phase 
+* This does not change actual order of column building - but adjusts which phase columns are built in 
 
 
 ### Version 0.3.1